% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelo_ARMA.R
\name{Analisis_series_temporales}
\alias{Analisis_series_temporales}
\title{Analisis de series temporales, modelo ARMA}
\usage{
Analisis_series_temporales(
  datos,
  p,
  q,
  n_predicciones,
  confianza = 0.95,
  graficar = TRUE
)
}
\arguments{
\item{datos}{debe de ser una serie temporal}

\item{p}{orden del modelo autorregresivo (AR)}

\item{q}{orden del modelo de media móvil (MA)}

\item{n_predicciones}{el numero de predicciones que se desea utilizar}

\item{confianza}{el nivel de confianza de los resultados}

\item{graficar}{si se desea ver grafica o no}
}
\value{
pronosticos futuros
}
\description{
El paquete resuelve una ecuación ARIMA (Autorregresión integrada de media móvil), que modela la serie temporal y permite hacer pronósticos futuros. La ecuación se ajusta a los datos de entrenamiento y utiliza el orden del AR (Autorregresión) y del MA (Media Móvil) especificados en la función. Luego, se realizan pronósticos con el modelo ajustado y se evalúa el rendimiento del mismo. Finalmente, se visualizan los resultados y se generan pronósticos futuros para un horizonte determinado.
}
\examples{
datos <- read.csv("ruta/archivo.csv", sep = ";")
print(datos)
serie_temporal <- ts(datos$toneladas, start = c(2022,1), frequency = 12)
print(serie_temporal)
datos <- Analisis_series_temporales(serie_temporal, 1, 2, 12)
}
